+++ fulcrum_trust/__init__.py
@@ -4,11 +4,12 @@ from fulcrum_trust.manager import TrustManager
 from fulcrum_trust.stores.file import FileStore
 from fulcrum_trust.stores.fulcrum import FulcrumStore
 from fulcrum_trust.stores.memory import MemoryStore
-from fulcrum_trust.types import TrustConfig, TrustOutcome, TrustState
+from fulcrum_trust.types import TrustCircuitOpen, TrustConfig, TrustOutcome, TrustState
 
 __version__ = "0.1.0"
 
 __all__ = [
+    "TrustCircuitOpen",
     "TrustManager",
     "TrustOutcome",
     "TrustState",